solution1

+++ app/main.py
@@ -9,17 +9,21 @@ def move_file(command: str) -> None:
         raise ValueError(f"Expected 3 arguements, {len(command_list)} "
                          f"were given!")
     if command_list[0] != "mv":
-        raise ValueError(f"Expected 'mv' as the first argument, "
-                         f"{command_list[0]} was given!")
+        raise ValueError(
+            f"Expected 'mv' as the first argument, "
+            f"{command_list[0]} was given!"
+        )
 
     source = command_list[1]
     destination = command_list[2]
+    if source == destination:
+        return
+
     dirname = os.path.dirname(destination)
 
     if dirname and not os.path.exists(dirname):
         os.makedirs(dirname)
 
-    with (open(source, "r") as orig_f,
-            open(destination, "w") as copy_f):
+    with open(source, "r") as orig_f, open(destination, "w") as copy_f:
         copy_f.write(orig_f.read())
     os.remove(source)
